backend/requirements-lambda.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
boto3==1.34.0
pymongo==4.6.1
passlib==1.7.4
google-auth==2.27.0
google-generativeai==0.3.2
